New utils functions, added datapack class

+++ opengl_current/IndexBuffer.hpp
@@ -11,36 +11,36 @@ class IndexBuffer
 {
 public:
     IndexBuffer(std::span<uint32_t> data, bool bDynamic = false);
-    IndexBuffer(int32_t maxNumIndices);
+    IndexBuffer(int maxNumIndices);
     ~IndexBuffer();
 
 public:
     void Bind() const;
     void Unbind() const;
-    int32_t GetNumIndices() const;
+    int GetNumIndices() const;
 
-    void UpdateIndices(const uint32_t* data, int32_t offset, int32_t size);
-    void UpdateIndices(const uint32_t* data, int32_t size);
+    void UpdateIndices(const uint32_t* data, int offset, int size);
+    void UpdateIndices(const uint32_t* data, int size);
 
     uint32_t GetOpenGlIdentifier() const;
 
-    static inline size_t s_IndexBufferMemoryAllocation{0};
+    static inline int64_t s_IndexBufferMemoryAllocation{0};
 
 private:
     uint32_t m_RendererId;
-    int32_t m_NumIndices;
+    int m_NumIndices;
 
 private:
     void GenerateRendererId(const uint32_t *indices, bool bDynamic);
 };
 
 
-FORCE_INLINE int32_t IndexBuffer::GetNumIndices() const
+FORCE_INLINE int IndexBuffer::GetNumIndices() const
 {
     return m_NumIndices;
 }
 
-FORCE_INLINE void IndexBuffer::UpdateIndices(const uint32_t* data, int32_t size)
+FORCE_INLINE void IndexBuffer::UpdateIndices(const uint32_t* data, int size)
 {
     UpdateIndices(data, 0, size);
 }
